chore: Move modern plugins to internal package
